etl-postgres-sync/requirements.txt
# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Data processing
pandas>=2.0.0

# Validation
pydantic>=2.0.0

# CLI
click>=8.0.0
rich>=13.0.0

# Scheduling
apscheduler>=3.10.0

# Logging
structlog>=23.0.0

# Retry logic
tenacity>=8.0.0

# Configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
